deal with sensor_state as normalized float

+++ dht22.h
@@ -1,9 +1,8 @@
 void dht22_setup();
 
 struct dht22_state {
-  // Stored with sensor-native scale: .1
-  int16_t temperature;
-  int16_t humidity;
+  float temperature;
+  float humidity;
 };
 
 struct dht22_state *dht22_read();
